Simplify the content hash check
The content hash currently has a somewhat complicated string of commands
so that the generation of it is predictable. This same command is being
used in preflight. I'd like to propose this new method, that instead of
sorting on the filenames themselves, this just saves the sort for the end
and sorts by the hashes.

Preflight would then be able to use a simpler set of code that is actual
golang instead of shelling out to a complicated command that is not very
testable on the preflight side.

Signed-off-by: Brad P. Crochet <brad@redhat.com>