Unify prod env naming
Ansible inventory uses 'prod' as a name of production environment and a
pipelines use 'production'. This is causing a failures when running a
pipeline in prod environment. Since we can't change inventory variable
because it determines ocp namespaces and few other things we have to
rename a 'production' -> 'prod' in a pipelines.