Change the URL for preflight.tar.gz.asc
Our need to potentially provide a docker config json required changing
our crypt step to encrypt and decrypt steps in the openshift-ci step
registry.

As a result of this change the job URL no longer has the word crypt but
instead encrypt. We build this URL dynamically however some parts of it
are hardcoded; this is one of those hardcoded parts.

Signed-off-by: Melvin Hillsman <mhillsma@redhat.com>