Temporarily fix issue with preflight job names
Currently we have to build 4.6 and 4.7 clusters from scratch and the
step to do this is different from claiming warm clusters from DPTP's
pool. Because the steps are different the job names are different;
preflight-common-aws and preflight-common-claim respectively. This is a
temporary fix until a better solution is found that addresses concerns
with building clusters from scratch versus claiming from the pool
as a whole.

Signed-off-by: Melvin Hillsman <mhillsma@redhat.com>