Create custom SCC with privilege escalation
The OpenShift pipelines 1.9 no longer allow running containers with
privilege escalation. Since a subset of pipeline tasks requires the
privilege escalation we need to use a custom SCC that is build on top of
pipeline SCC and add the escalation.

This commit adds the custom SCC, links it with the pipeline service
account and refactors ansible to be compatible with the new setup.

JIRA: ISV-3280