Show pipeline status using PR labels
The new Python module has been added to manage labels in Github PRs. The
module allows adding and removing labels.

The new task is introduced which sets labels for a PR. The task is used
in all pipelines and indicates a status of the pipeline by setting
appropriate label.

JIRA: ISV-3895