Move signing pipeline into separate namespace
The index image signing pipeline is now part of a separate namespace and
it is not exposed using nginx proxy.

The need for having external route
is no longer valid since the only external consumer was community
pipeline. With the new community pipeline signing solution is already
embedded. The signing pipeline in a separate namespace will be only used
by cert ops.

JIRA: ISV-3501

Signed-off-by: Ales Raszka <araszka@redhat.com>