Ask for a review using comments
The previous solution by adding usernames as a PR reviewers didn't work
if users were not part of the organization. The previous command
returned 404 error.

We switched to a asking for review using comments. A comment is added
and users from ci.yaml are tagged in the comment.

JIRA: ISV-4381

Signed-off-by: Ales Raszka <araszka@redhat.com>