Publish resources using image requests
The publishing process of metadata in Pyxis is handled by unified
process that is common for both containers and operators.

The process required a bunch of refactoring and simplification in the
pipeline. This commit removes all unused or duplicated steps and reduce
pyxis interaction to only one task that handles everything.

JIRA: ISV-3989

Signed-off-by: Ales Raszka <araszka@redhat.com>