Skip preflight test only when previous attempt was sucesfull
There was a bug introduced in recent commit that cause a preflight to be
skipped even for previous attempts that weren't successful.

This commit fixes the workflow to skip a preflight only when test was
successful and always execute preflight when previous attempt wasn't
found or attempt failed.

JIRA: ISV-4741

Signed-off-by: Ales Raszka <araszka@redhat.com>