Add user facing documentation and mkdoc
The mkdoc is used to generate a dynamic pages with a user documentation.
The documentation is common for ISV and community operators. The docs
will server as a official doc for operator users.

The mkdoc will run whenever a doc is updated and will be released to
Github hosted environment.

JIRA: ISV-4507

Signed-off-by: Ales Raszka <araszka@redhat.com>