Enable bundle removal for FBC operators
A user can remove a bundle from repository of it uses FCB workflow and
the removed bundle is not used in any of existing catalogs. A user can
also combine a bundle removal and catalog updates so they can remove
entire bundle and catalog at the same time.

JIRA: ISV-5138

Signed-off-by: Ales Raszka <araszka@redhat.com>