Migrate to a new basic template schema
A opm introduced a new breaking change and changed a basic template
schema. The schema produced by the fbc migration tool was no longer
supported and required a change.

This commit adapts to a new opm version and produce new basic template
schema. It also updates a onboarding doc to reflect a changes and does
couple of user improvements.

JIRA: ISV-5164

Signed-off-by: Ales Raszka <araszka@redhat.com>