Add script for auto-releasing bundle to FBC
The new script is introduced that automatically release a new bundle to
a fbc template and renders it to catalog using opm command.

The script creates templates if not exists yet or amend template and add
new bundle. The script uses a new catalog mapping in ci.yaml and new
release-config.yaml file that contains info about how to release a
bundle.

JIRA: ISV-5506

Signed-off-by: Ales Raszka <araszka@redhat.com>