fix: Skip operator test in case it gets removed from repository
Static check used to fail on PRs that removed the resource in a PR. This
is not an intended behaviour and pipeline should not fail for such a
PRs.

This commit extends a logic how resources for static checks are
collected and check their presence before submitting them to test suite.

JIRA: ISV-5515

Signed-off-by: Ales Raszka <araszka@redhat.com>