Auto-release bundle to FBC
A new steps is added to hosted and release pipeline that is executed if
a user submits a fbc bundle with release config.

Base on the release config a pipeline inject the new bundle into a
catalog templates and renders a final catalog from it.
In the hosted pipeline we only run a dryrun to make sure the catalog is
still valid after the release. In the release pipeline we also commit
the changes and open a new PR with the changes. The new PR is handled by
the separate pipeline that releases the FBC.

In order to be able commit and open a PR a pipelines need to be changed
to use ssh git pull URL. A new ssh key has been released that allows
pushing new changes to a upstream. The SSH key is added to the bot
account.

JIRA: ISV-5507

Signed-off-by: Ales Raszka <araszka@redhat.com>