Add Tekton-lint to tox configuration
Tekton lint validates Tekton tasks and pipeline and suggest
improvements. This commit contains a basic configuration of the linter
in the tox and also fixes all violations.

The linter is not enable by the default because it crashes on
definitions that uses a tekton bundles. Due to this restriction it only
serves as optional tool.

JIRA: ISV-5288

Signed-off-by: Ales Raszka <araszka@redhat.com>