ISV-5645: Signing pipeline for image bootstrapping
A new pipeline is created that will server for the automating of index
image bootstrapping process. The new pipeline allow signing only images
that haven't been configured in Pyxis as GA.

The new pipeline will be deployed into separate namespace which will be
restricted only to a service account that can trigger a pipeline.

JIRA: ISV-6546

Signed-off-by: Ales Raszka <araszka@redhat.com>