fix: add-bundle-to-fbc uses registry credentials
In order to render a catalog from template a task needs a registy
credentials. The auth file was mounted as a workspace to the task and
default env variable is pointing to the auth config.

While testing the config I found out that pre-prod registry tokens lost
access to registry and needed to be rotated.

JIRA: ISV-5757

Signed-off-by: Ales Raszka <araszka@redhat.com>