Skip preflight-trigger based on PR label
A PR label can now skip entire preflight-trigger step.

This feature is useful especially for ack-* operators that don't depend
on the preflight checks but their PRs still triggers the job and
consumes resources. Instead of providing a real preflight output we mock
the result and provide an empty values.

JIRA: ISV-5988

Signed-off-by: Ales Raszka <araszka@redhat.com>