Add code for webhook dispatcher
A dispatcher is a new component to manage incoming webhooks from Github.
It accept new requests and manage a queue based on cluster capacity.

A capacity is defined as a number of running pipelines in certain
namespace. This simple metrics will provide a buffer mechanism for time
periods when we receive a bunk releases of operators. The mechanism
keeps pending requests in the database queue and only trigger related
pipeline in case of free capacity.

The solution is made of:
- Rest API
- Postgres database
- Dispatcher
- Capacity manager

JIRA: ISV-6108

Signed-off-by: Ales Raszka <araszka@redhat.com>