Deployment for webhook dispatcher app
This commit brings a deployment configuration for a new webhook
dispatcher application.

It consist of following parts:
 - postgres database
 - webhook dispatcher application
 - services, routes, secrets, pvc, configs

The application deployment is fully driven by the Ansible.

As part of the commit I also split the "local" user workspace into
separate ansible inventory that share a data with stage but can be
customized.

JIRA: ISV-6109

Signed-off-by: Ales Raszka <araszka@redhat.com>