Add retry mechanism for registry push commands
There are 2 types of commands that are handled by this commit.

1. Bash commands executed in the Tekton task
2. Commands executed in the Python script

Both of these uses re-tries mechanism now. The bash commands use a
generic script that controls the execution and retries if needed.

This commit "fixes" the tasks and commands that failed the most often
and improves stability.

JIRA: ISV-6114

Signed-off-by: Ales Raszka <araszka@redhat.com>