fix: Re-arrange ansible inventory groups
As a side effect to previous commits that introduced a playground group
I unintentionally caused that both stage and playground have been
deployed when running stage release.

This commit re-arrange groups in a way that both share same parent group
to make sure we share same group vars but both groups are independent.

Signed-off-by: Ales Raszka <araszka@redhat.com>