Wait for db and dispatcher to be operational
In the integration test we can see that dispatcher and underlying db is
not ready at the time when tests are executed. This commit adds a wait
steps to wait for db and dispatcher.

It also adds an local dev environment guide with how to setup a
environment using dispatcher.

The commit also increase a default capacity to 5 PRs instead of 3.

Signed-off-by: Ales Raszka <araszka@redhat.com>