Add queued label to PR for events in queue
A webhook dispatcher sets a PR label once there is not enough capacity
on a cluster and event gets in a queue.

The queue label matches a current label pattern for each pipeline. The
example for hosted pipeline is "operator-hosted-pipeline/queued". This
format makes sure the label is removed once the pipeline starts
processing the PR.

The label is set only once when the event gets in a queue. The event
corresponding DB entry is also changed to indicate queued status.

JIRA: ISV-6265

Signed-off-by: Ales Raszka <araszka@redhat.com>