fix: Merge task fails with automerge
Due to what seems to be like a bug in github CLI or directly in Github
the --auto is now replaced with --admin.

The auto flag sometimes raised a Graphql error from which it was hard to
tell what's wrong. The issue was pretty non-deterministic and pretty
much failed randomly. A local testing didn't prove the issue on our
site. This issue was most likely introduced by start using the Github
CLI utility. Previously the merge was done using direct API calls to
REST. The current CLI tool uses GraphQL.

At the point when pipelines wants to merge the PR we already know
whether a test passed or failed so it is safe to use --admin and merge
the PR immediately.

JIRA: ISV-6111

Signed-off-by: Ales Raszka <araszka@redhat.com>