feat: Build fragment images with only changed content
Previously we always added all FBC operators to fragment image. This
however showed a downside when pipeline is manually restarted. Due to a
restart the pipeline can include content to fragment that is outdated or
even removed.

This change detects which catalog operators have been updated and only
add these to the fragment. It reduces a image size and avoid race
condition when pipeline is manually rebooted.

JIRA: ISV-6459

Signed-off-by: Ales Raszka <araszka@redhat.com>